Mark unsafe functions as unsafe.
Basically I looked at #142 and said "wait, if passing the wrong
parameters to this function can cause memory corruption, and
there's no good way we can automatically add checks for that,
this function should be unsafe".

So I added `unsafe` to everywhere in `EncoderCommon` that it looked
like it needed it, generally 'cause it called an unsafe `gfx-hal`
function and passed parameters through without doing any bounds
checking.  Then I added unsafe blocks where necessary to make the
thing compile.  This PR is offered as a matter for discussion
and feedback to figure out how we want to handle this issue;
the code is far from complete.