Raw string source shader support (#141)
Merge #141 

141: Raw string source shader support r=Frizi,omni-viral a=jaynus

Deprecates StaticShaderInfo in favor of PathBufShaderInfo. Implements SourceShaderInfo publically which allows for providing raw GLSL source code to pass to shaderc-rs. Adds new example source_shaders for usage. All examples updated for deprecated functionality.

This move soft-deprecates the current StaticShaderInfo which actually wasn't in line with the current implementation; this was supposed to support string shader info to begin with, but instead just provided a path. This instead deprecates that access, and pushes all implementations to PathBufShaderInfo, which enforces a proper PathBuf be supplied (this will actually fix support for non-english user filesystems better).

Then, this changes the current SourceShaderInfo to FileShaderInfo, to better reflect its implementation. Finally, a new implementation SourceCodeShaderInfo is made, which loads from source, and then is aliased as SourceShaderInfo for public consumption.

Closes #136
Closes #140 

Co-authored-by: Walter Pearce <jaynus@gmail.com>