Add methods and types for reading from `MeshBuilder`
MeshBuilder is a convenient way of storing mesh data on the CPU, since it supports a
wide variety of formats due to using arbitrary vertex attributes. However it does
not have any way to read the data stored inside of it.

This patch adds functionality for reading the mesh stored inside of the buffers:

* `MeshBuilder::view_attr` finds an attribute in the stored buffers by name and returns a view
  into the buffer that reads that attribute. It can also be iterated over.
* `MeshBuffer::iter_index` iterates over elements in the index buffer, or from 0..numvertices if there
  is no index buffer.
* Various getters for `MeshBuffer` and `RawVertexBuffer`
* `FromVertexBuffer` defines how to read a type from a vertex buffer and convert it from the vertex
  format its in (ex. normalizing and scaling integer values into floating-point).